+++ pkg/cmd/login.go
@@ -38,5 +38,5 @@ func (lc *loginCmd) runLoginCmd(cmd *cobra.Command, args []string) error {
 		return login.InteractiveLogin(cmd.Context(), &Config)
 	}
 
-	return login.Login(cmd.Context(), lc.dashboardBaseURL, &Config, login.AsyncStdinReader{})
+	return login.Login(cmd.Context(), lc.dashboardBaseURL, &Config)
 }
