+++ pkg/login/interactive_login.go
@@ -15,6 +15,7 @@ import (
 
 	"github.com/stripe/stripe-cli/pkg/ansi"
 	"github.com/stripe/stripe-cli/pkg/config"
+	"github.com/stripe/stripe-cli/pkg/login/acct"
 	"github.com/stripe/stripe-cli/pkg/stripe"
 	"github.com/stripe/stripe-cli/pkg/validators"
 )
@@ -51,10 +52,10 @@ func InteractiveLogin(ctx context.Context, config *config.Config) error {
 }
 
 // getDisplayName returns the display name for a successfully authenticated user
-func getDisplayName(ctx context.Context, account *Account, baseURL string, apiKey string) (string, error) {
+func getDisplayName(ctx context.Context, account *acct.Account, baseURL string, apiKey string) (string, error) {
 	// Account will be nil if user did interactive login
 	if account == nil {
-		acc, err := GetUserAccount(ctx, baseURL, apiKey)
+		acc, err := acct.GetUserAccount(ctx, baseURL, apiKey)
 		if err != nil {
 			return "", err
 		}
