+++ pkg/login/login_message.go
@@ -6,13 +6,14 @@ import (
 	"os"
 
 	"github.com/stripe/stripe-cli/pkg/ansi"
+	"github.com/stripe/stripe-cli/pkg/login/acct"
 )
 
 // SuccessMessage returns the display message for a successfully authenticated user
-func SuccessMessage(ctx context.Context, account *Account, baseURL string, apiKey string) (string, error) {
+func SuccessMessage(ctx context.Context, account *acct.Account, baseURL string, apiKey string) (string, error) {
 	// Account will be nil if user did interactive login
 	if account == nil {
-		acc, err := GetUserAccount(ctx, baseURL, apiKey)
+		acc, err := acct.GetUserAccount(ctx, baseURL, apiKey)
 		if err != nil {
 			return "", err
 		}
